crystallizer-3 create and implement granular delay logic
granular delay logic working, implemented an update function to keep the granular delay parameters up to date with the sliders. matched the parameter names correctly to their sliders, learned that registering parameters and setting up the update functions is unfortunately Quite Annoying. considering a refactor in this area

Modified the pitch shift so that it operates on semitones